ipcpd: Fix shutdown of link-state routing
The link_state component tried to unregister from the notifier while
it had threads running.

The deadlock happens when notifier_event had a rdlock trying to delete
a neighbour, while notifier_unreg tries to take the write lock. And
the delete can't complete in the notifier.

https://tree.taiga.io/project/dstaesse-ouroboros/us/113

We need to make sure that all threads are stopped before unregistering
from the notifier.

Updated link_state to use the start/stop() paradigm and not
create/cancel threads within init/fini().

Signed-off-by: Dimitri Staessens <dimitri@ouroboros.rocks>
Signed-off-by: Sander Vrijders <sander@ouroboros.rocks>